refactor: unify classify_llm_error into single canonical definition
- Move classify_llm_error and is_context_window_error to astra_core::error_kind
- Delete duplicate copies in client.rs, turn_ingest.rs, and inprocess.rs
- Fix bug: turn_ingest.rs was missing pool timeout → DatabaseError check,
  causing pool timeout errors to be misclassified as StreamIdle
- Add 9 focused tests for classify_llm_error and is_context_window_error
- Net -80 lines, zero semantic changes beyond the bug fix
